requirements/base.txt
pytz==2022.4  # https://github.com/stub42/pytz
python-slugify==6.1.2  # https://github.com/un33k/python-slugify
Pillow==9.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
redis==4.3.4  # https://github.com/redis/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py
celery==5.2.7  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.5.0  # https://github.com/celery/django-celery-beat
flower==1.2.0  # https://github.com/mher/flower
requests-aws4auth==1.1.2 # https://github.com/tedder/requests-aws4auth
python-json-logger==2.0.7 # https://github.com/madzak/python-json-logger
django-extensions==3.2.3 # https://github.com/django-extensions/django-extensions

# Django
# ------------------------------------------------------------------------------
django==4.2.8  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.9.0  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.51.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.4.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==3.13.0 # https://github.com/adamchainz/django-cors-headers
djangorestframework-simplejwt==5.2.1 # https://github.com/jazzband/djangorestframework-simplejwt
# DRF-spectacular for api documentation
drf-spectacular==0.27.0  # https://github.com/tfranzel/drf-spectacular
djangorestframework-camel-case==1.4.2 # https://github.com/vbabiy/djangorestframework-camel-case
django-orderable
python-telegram-bot
